+++ RELEASE_NOTES.md
@@ -29,6 +29,7 @@ This serves two purposes:
 - Added missing collection key types in Hyde facade method annotations in https://github.com/hydephp/develop/pull/1784
 - Fixed heading permalinks button text showing in Google Search previews https://github.com/hydephp/develop/issues/1801 in https://github.com/hydephp/develop/pull/1803
 - Realtime Compiler: Updated the exception handler to match HTTP exception codes when sending error responses in https://github.com/hydephp/develop/pull/1853
+- Realtime Compiler: Improved routing for nested index pages in https://github.com/hydephp/develop/pull/1852
 
 ### Security
 - in case of vulnerabilities.